Add Alex Trudel's Work
Sorry for this being so late, but I had some issues with the code not doing what I wanted it to. Overall, the code is really inefficient but I just couldn't find anything that could work out. It's probably an easy fix. Also the analysis isn't quite done but I'll be working on it first thing tomorrow, then I'll commit the new version.